MandelbrotRunner as interface introduced
